ARM: dts: Add MSM7X30 support [WIP]
MSM7X30 is not the correct name according to DT specification, rather it should
be named msm7230 or msm7630 (more preferred). msm8255 is also possibility.

Pinctl gpios should not be done in the qcom-msm7x30.dtsi, rather it should be in
msm7630-surf.dts or similar.

Also defconfig needs checking if it survived the update properly.

Change-Id: I4da1d89f038751bf5aa9e8883dc9b91d231cc7a0