clk: qcom: Initial ACPU clock driver for msm7x30
This is initally imported from the LA.AF.1.1_rb1.17, however it is a
complete makeover to make the ACPU clock a generic clock that a CPUFREQ
driver can use normally. This incorporates AXI frequency voting.

WIP: The parent clocks that ACPU can source from should be loaded from DT. But
how, if we need to vote for a "not sourced" AXI clock?

Change-Id: I45030d67d3121faac80a10d75c6c5e345ad690dd