leds: as3676: Add DT support
Other changes include removal of header as all users of this driver
are using DT. Updated kernel APIs for strict_strtoul->kstrtoul and use
of devm_* API for memory allocation. ALS config is currently not read
from DT, as no device has requested doing it. If such need comes in the
future, the driver must be adapted.

Change-Id: Iee9d93a9ecf18b69e8f9bf204315309dacdf7811